libhsakmt/virtio: fix 32-bit truncation of vhsakmt_bo.size (>=4GB allocs)
struct vhsakmt_bo.size was declared "unsigned" (32-bit). A device
allocation of exactly 4 GiB (0x100000000) truncated to 0 when stored,
so vhsakmt_bo_cpu_map() -> virtio_gpu_map_handle() called mmap(len=0),
which returns EINVAL. ROCr then reported
"MemoryRegion::BlockAllocator::alloc failed / out of memory" for any
single >=4GB device buffer (e.g. a llama.cpp 128K-context KV cache).
Widen the field to uint64_t so the full size reaches mmap().

Proven via strace on the guest: mmap(0x..., 0, ...) = -1 EINVAL at 4GB,
vs a correct-length mmap at <4GB. Verified 2GB allocs still succeed.

NOTE: this unblocks the mmap, but mapping a single 4GB HOST3D blob
host->guest still stresses the Xen HMEM/p2m path; large device blobs
should additionally be chunked to a safe size (follow-up).

+++ projects/rocr-runtime/libhsakmt/src/virtio/hsakmt_virtio_device.h
@@ -126,7 +126,7 @@ struct vhsakmt_bo {
   struct vhsakmt_device* dev;
 
   _Atomic int refcount;
-  unsigned size;
+  uint64_t size;   /* was unsigned (32-bit): truncated >=4GB allocs -> mmap(0)=EINVAL */
   void* cpu_addr;
   void* host_addr;
   HsaMemFlags flags;
